daemon/c_seccomp: Moved mknod emulation code to seperate file
This moves the emulation code for mknod() and mknodat() syscalls
into the separate 'c_seccomp/mknod.c' file. We did not change the
emulation code itself. Just moved c_seccomp_emulate_mknodat()
to the new file and generated necessary interface code.
